Rename ICalGenerator to EventToICal
