+++ app/logic/Calendar/Event.ts
@@ -221,7 +221,7 @@ export class Event extends Observable {
   /** DTSTAMP when the status was sent that is captured in this object.
    * Used during auto-update to avoid overwriting with older info.
    * We don't need to track local changes, only those sent by others.
-   * Local changes sent to others will always get the current timestamp in `ICalGenerator` */
+   * Local changes sent to others will always get the current timestamp in `EventToICal` */
   lastUpdateTime: Date | null;
   /** Includes changes to `alarm`, lastUpdateTime does not consider to be a change */
   @notifyChangedProperty
